requirements.txt
fastapi
uvicorn
sqlalchemy 
passlib[bcrypt]
PyJWT
python-multipart